requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
edge-tts==6.1.9
faster-whisper==1.0.3
moviepy==1.0.3
pexels-api==1.0.1
google-generativeai==0.3.1
playwright==1.40.0
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
